requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph
chromadb
sentence-transformers
numpy
python-dotenv
ipython
re
python-multipart 
twilio
fastapi
uvicorn